Add files via upload
This product has improved the chessboard function on the basis of version 1.0, and can correctly determine the outcome according to the general rules of backgammon. It has also made preliminary construction for version 2.0 and set up a basic version of AI, which can compete with players with the goal of victory.